+++ kotlin/src/main/kotlin/aoc2019/intcode/instructions/control/JUMP_IF_TRUE.kt
@@ -8,8 +8,8 @@ object JUMP_IF_TRUE: Instruction2 {
 
     @Suppress("PARAMETER_NAME_CHANGED_ON_OVERRIDE")
     override suspend fun State.exec(condition: Param, destination: Param): Boolean =
-        if (read(condition) != 0) {
-            setPointer(read(destination))
+        if (read(condition) != 0L) {
+            setPointer(read(destination).toInt())
             true
         } else false
