+++ kotlin/src/main/kotlin/aoc2019/intcode/io/DefaultIO.kt
@@ -1,12 +1,12 @@
 package aoc2019.intcode.io
 
-open class DefaultIO(settings: Array<Int> = emptyArray()) : IO(settings) {
-    override suspend fun provide(): Int {
+open class DefaultIO(settings: Array<Long> = emptyArray()) : IO(settings) {
+    override suspend fun provide(): Long {
         print("Please enter the input: ")
-        return readln().toInt()
+        return readln().toLong()
     }
 
-    override suspend fun write(out: Int) {
+    override suspend fun write(out: Long) {
         println(out)
     }
 }
